Tweak healthcheck response body
We want to give some simple static response here, but {"200": "OK"}
isn't ideal. The response status is meant to be communicated in the
HTTP status code. Embedding in the response body as well encourages
callers to wrongly ignore the HTTP code and opens the possibility
that the two codes might not match.

Let's use the same response model as FastAPI uses by default for
errors, which is:

- no embedded response code
- a "detail" attribute with human-oriented text

This isn't a big deal since this is mainly for internal use, but
let's get it right before adding other APIs.