scripts: add call-context helper for dealing with auth
The call context header used for authentication is base64 encoded,
so it is not practical to generate header values manually, which is
annoying when trying to test the API with tools such as curl.

Add a tiny helper script for development to improve testability.
It can be used with curl as in example:

    curl \
      -H "X-RhApiPlatform-CallContext: $(scripts/call-context some-role)" \
      http://localhost:8080/some/url