Make dependabot use Python 3.6 while resolving deps
Dependabot is currently permitted to use the latest available
Python version. This doesn't work for us - we need pip-compile to
be run using Python 3.6, same as used in our container images,
otherwise the dependencies can be resolved incompatibly.

There is no dependabot-specific config file to control this. It
instead reads the configuration of various other python tools, and
there are several options for setting the version.
Adding it to pyproject.toml in this way seems the best option, as
we can have an explanatory comment and it does not block
development & testing of the project on later python versions.