Implement environment support in upload API [RHELDST-2188]
This commit removes the 'bucket' parameter from the s3 upload API and
it's endpoints, replacing it with 'env'. The bucket name and other
environment-specific configuration is now drawn from the corresponding
environment defined in exodus-gw.ini.

Additionally, the s3_client method now takes an AWS profile name to
specifiy the AWS configuration desired for the resulting client.