Make dependabot use Python 3.6 while resolving deps via runtime.txt
This is a second attempt at making dependabot use Python 3.6 while
running pip-compile, since this is our target environment and using
the same Python version is necessary for updates to requirements
files to work properly.

runtime.txt originally came from Heroku, which we don't use, but
it allows dependabot to pick the correct Python version without
affecting anything else.  Compared to the previous approach of
using pyproject.toml, it has a disadvantage that the file format
doesn't allow for comments.

See commit c430043b6884 for the first attempt.