deps: require uvicorn[standard] rather than just uvicorn
The reason for this is, as mentioned in the uvicorn 0.12.0 changelog:

> Uvicorn no longer ships extra dependencies uvloop, websockets
> and httptools as default. To install these dependencies use
> uvicorn[standard].

If those deps are not installed, that would effectively mean that
upgrading to uvicorn 0.12.0 switches us from (uvloop & httptools)
backends to (asyncio & h11) backends.

uvicorn docs[1] imply that the former are superior, so let's tweak
the deps to ensure we keep using them.

[1] https://www.uvicorn.org/settings/