Introduce dev env helpers using systemd & podman
This commit adds some optional helpers which allow running
exodus-gw and supporting services locally within systemd user
units, along with docs on how to use it.

The included components are currently:

- exodus-gw web app (uvicorn)
- sidecar proxy, for auth/https
- postgres server
- localstack (fake s3)

It's enough to develop and test upload features in exodus-gw
locally with no dependency on remote services. The localstack
instance hopefully will provide a working fake dynamodb as well,
but this hasn't been tested or enabled yet.

To get started trying it:

- you should use a modern version of Fedora
- then run scripts/systemd/install and follow instructions