Avoid confusing UnseekableStreamError on request retries [RHELDST-3407]
This exception was observed from exodus-gw logs when trying to use the
upload API:

  botocore.exceptions.UnseekableStreamError: Need to rewind the stream
  <exodus_gw.s3.util.RequestReader object at 0x7f1869ebe9e8>, but stream is not seekable.

The reason for this is the combination of the following:

- For best performance and lowest overhead, as we receive requests,
  we stream the request body to S3 without saving it in memory.

- This means there is no way to read the request body more than once,
  as we haven't stored it anywhere.

- The boto library has a couple of features which cause it to attempt
  reading the request body more than once. One is retrying failed
  requests, which is documented, configurable & enabled by default.
  Another is automatically redirecting you to the correct region if
  you set up the s3 client pointing at the wrong region.

Altogether this means that, if a request to S3 failed for some reason
or was pointed at the incorrect region, the original error may be
swallowed and replaced by this UnseekableStreamError as boto attempts
a second request.

To fix that, let's disable the functionality in the boto library
which can cause it to retry requests.

This is a good idea anyway, as the exodus-gw client will already
have retry logic by default and this should not be doubled up in
both places; and because the implicit region redirect is an
antifeature for us, as we'd rather detect a wrong region in config
immediately.