Fix batch_write put/delete requests
Previously, the put/delete request contructor attempted to duplicate the
"Item" key. Since duplicating keys is not permitted in Python, this
resulted in the submission of only the last item from the given list.

This change corrects this bug, creating put/delete request dicts for
every item in the given list.