Implement commit_publish endpoint [RHELDST-2161]
This commit introduces a new endpoint used to commit the items stored on
existing publish objects to their place in the given environment's DynamoDB
table. Retry with exponential backoff and jitter was added to the
batch_write function.

The endpoint finds the publish based on the given publish_id,
divides the publish's items into appropriate batch sizes and executes
batch_write operations. Upon failure, the entire operation is aborted
and an attempt is made to delete what was written.