devenv: tweak uvicorn args to avoid unnecessary reloads
The intent here is to reload automatically when code has changed,
so let's have it look under the directory containing the code only.

Without this, it'll look at '.', which means it'll frequently reload
as irrelevant files are changed (e.g. it'll watch files under .tox
and then reload if you run any tox command).