cleanup: drop unnecessary constructors from models
One of the features included in DeclarativeBase is the implicit
generation of constructors with keywords matching attributes on
the models.  We don't need to define constructors ourselves, and
it's probably best to avoid doing that since it could lead to
inconsistencies such as forgetting an attribute in one place or
the other.