dev env: support usage of dynamodb in localstack
The dynamodb service in localstack seems to work just fine, so
let's enable it and start making use of it.

It's necessary to create a table before this will work. Because
the command for that is somewhat cumbersome, it was wrapped up
in a helper script; the same thing was done for s3 for consistency.

Docs now contain a few additional commands which can be used to
check the state of s3 & dynamodb within localstack.