Introduce 'exodus-sync' example script
This example covers a full recursive sync workflow. It works
similarly to a recursive scp or rsync, but by making use of the
exodus-gw APIs.

This can be used as a quick way to exercise the upload and publish
APIs in a development environment. With latest patches, the entire
process is basically working, but there are various TODOs to be
solved later due to some bugs and unimplemented features in
exodus-gw.