Refactor lifecycle of DB session, commit if request succeeds
It's a common pattern in web apps to have each incoming request
wrapped in a transaction, which is committed at the end of the
request if the request handler succeeds, or rolled back if it
fails.

Let's do it in exodus-gw as well, which avoids most code having to
explicitly commit (which can be easily forgotten, leading to bugs).

This is largely motivated by upcoming changes to the usage of
dramatiq. It's intended to make dramatiq and sqlalchemy share the
same transaction, but once this is done, it means that any usage
of a dramatiq actor must be combined with a COMMIT. Since dramatiq's
usage of a DB is hidden, and the caller might not even hold a
reference to the current SessionLocal, this also could easily be
forgotten - better to do it implicitly to rule out this class of
bugs.

In order to make this work, some refactoring is needed for the
lifecycle of DB sessions. DB sessions were previously created
and closed from within a dependency ("Depends"). Unfortunately,
as explained in [1], the dependency function won't be able to see
whether an exception has occurred, so it doesn't know whether it's
appropriate to commit.

To resolve this, DB session handling was moved into middleware.
The DB session can still be obtained via Depends(get_db).

Additionally, there were some developments in the test suite to
make this testable meaningfully. sqlalchemy is now bound to an sqlite
DB while tests run. This means tests can be written using the FastAPI
TestClient to exercise HTTP and DB handling "for real", where
appropriate.

A couple of mock-heavy tests were removed as they seem to be
obsoleted by having at least one test which genuinely uses a DB.

[1] https://fastapi.tiangolo.com/tutorial/dependencies/dependencies-with-yield/#dependencies-with-yield-and-httpexception