Greatly extend the docs available in the API reference, drop API guide
API docs were previously separated into two pieces:

- API user guide: general usage information, and info on topics
  covering all APIs such as authentication
- API reference: detailed reference of each individual endpoint

It was mainly set up this way because it wasn't clear to me how to
integrate larger chunks of docs into the openapi-generated docs
used for the reference.

It turns out that it's possible to have any desired docs appear
in the reference by using the appropriate parameters when
configuring our FastAPI app and routers.

It's also possible to have docs associated with each openapi tag,
and to group docs under headings and link to them using markdown.
I think this is enough to obsolete the separate API user guide
and keep all the docs in one place, so let's do that instead.

This commit includes:

- explicitly defining our openapi tags so we can associate
  meaningful docs with them

- moving all contents of the API user guide into the appropriate
  openapi docs

- filling in some missing docs, particularly for the publish APIs

Note: the docs sometimes refer to behavior we intend to implement
rather than what is currently implemented.