Fix LookupError in worker introduced by 6c7b03200
Commit 6c7b03200 made sqlalchemy and dramatiq share DB sessions.

To make this asyncio-safe, a contextvar was used to hold the current
session during an HTTP request. However, the implementation broke
the background worker process, which would raise an error like this
on startup:

  LookupError: <ContextVar name='pool' at 0x7fe3effd2630>

My mistake was thinking that the current context is swapped only
when switching coros and not when switching threads. Turns out
it's swapped at both, which means that the dramatiq worker would
lose its reference to the pool when booting up new worker threads.

To fix this, make the broker use a ContextVar only after it has
been asked to use an sqlalchemy session. This makes the overall
behavior as follows:

- within the web component, the broker uses a ContextVar so that
  each separate HTTP request, which may be running concurrently,
  keeps its own pool pointing at its own session. It is not shared
  by any other thread or coro.

- within the worker component, the broker uses the default pool
  as created by dramatiq-pg. This is a ThreadedConnectionPool
  which is safely shared by all threads.

It's rather messy that our broker class now has these two different
modes of operation. I played around with setting a different broker
class for the web and worker components, but unfortunately as the
design of dramatiq strongly encourages the broker to be fully
configured at import time, that approach also seems to be quite
messy.