refactor: obtain Environment object via dependency system
Where an endpoint needs an Environment object, we can obtain it
via deps.env rather than calling get_environment from within the
endpoint.

The motivation for this is to enable some subsequent refactors to
get_environment and Settings. If we obtain environment through the
dependency system, then the signature of get_environment can later
be changed without requiring further changes from the endpoints.

Note a handful of test changes due to this:

- a couple tests were modified to use TestClient rather than invoking
  endpoint function directly. This shows that the usage of
  Depends(get_environment) works correctly (e.g. it can access 'env'
  path parameter).

- a few tests of 'access nonexistent env' were removed, since the
  usage of Depends means the endpoint functions themselves can't
  possibly be invoked with a nonexistent env any more

- other tests were simply modified to pass an Environment object
  rather than a str.