refactor: obtain Settings object via dependency system
Previously, Settings was expected to be loaded via a get_settings()
function which used a global shared cache. Since settings were used
at import time, that meant we would load settings once at import
time and never again.

This has a negative impact on testability, so let's refactor the
lifecycle. Settings are now initialized during the FastAPI startup
event and cached for the lifetime of the app, and declared as a dep
where needed.

The advantages of this are:

- for low-level tests, if you want to test what's the behavior of
  a certain function when given certain settings, you can simply call
  it with the desired values. No need to attempt mocking get_settings
  or messing with the cache.

- for high-level tests, if you monkeypatch.setenv("EXODUS_GW_*", ...)
  to control settings before constructing TestClient in your test, it
  will work as expected. This is unlike the previous state where the
  settings would have already been initialized at import time with no
  way to reinitialize them.

Note this commit doesn't solve all problems relating to settings; in
particular, the DB engine initialization still uses settings initialized
at import time only, and it's unclear when settings should be loaded
within dramatiq actors. There will be more refactors in upcoming commits.