refactor: create DB engine during startup, not import
Previously we had defined SQLALCHEMY_DATABASE_URL, the DB engine
and session factory at import time.

Although this is simple, it has a negative impact on testability
of the code. It implies that Settings must be evaluated at import
time, which means writing a test which adjusts settings is not
going to work consistently (e.g. the value of
SQLALCHEMY_DATABASE_URL wouldn't be recalculated as settings
change).

Let's adjust the lifecycle so that the DB engine is constructed
during app startup. This allows tests to point at a different DB
just by adjusting environment variables, as expected.