fixup: drop redundant constraints on primary key columns
It's not necessary to set unique or non-null constraints on
primary keys, because those are always unique and non-null.

This has been harmless, but this should be cleaned up for
alembic migrations to work more cleanly - otherwise, alembic
autogenerate will always detect that some uniqueness
constraints need to be created.