docs: fix most settings missing from deployment guide
We had disabled the usual behavior of listing all members here.
I think this was done because 'environments' is computed and can't
be set directly by env var, but this way it's too easy to forget
adding a useful setting to docs.

Let's just use the convention that we will include all documented
members in the docs, and simply don't put doc strings on those we
don't want to appear.