ci: start doing type checks with mypy
"tox -e static" now includes type checks using the mypy tool.
It will detect some kinds of type errors.
This commit also includes fixes to errors currently detected by
the tool (all of which are minor).

The motivation here is that there have been a handful of errors
introduced which could have been prevented by these checks.
One example is in refactor a64bb00b7e, where the 'env' parameter
to some functions was changed from str to Environment, but this
was not done properly everywhere (and tests didn't pick it up).

Note the tool is using a very tolerant config: it's not mandatory
to add type info anywhere, and it's also not mandatory that type
info is available for all the libraries we're using. Based on
experiences in other projects using mypy, I would rather not try
to enable the stricter options.

The effectiveness of the tool at each part of the code will depend
on how much type info a developer chooses to add.