Add get_task endpoint [RHELDST-2162]
This commit introduces a new endpoint that returns a task object from
the database using the given task ID.

Database schema updated to include a "tasks" table, which is used to
update and track the status of dramatiq worker tasks.

The "commit" actor, responsible for delegating writes to DynamoDB in
dramatiq workers, was modified to update the task state as it
progresses.