refactor: move dramatiq initialization to middleware
Previously the dramatiq schema initialization was hooked into
'consume' on the broker.

It turns out this wasn't the best place to put it because this is
called once per queue per thread, so we'd run the initialization
code many more times than necessary, and in parallel.

We should be using middleware for this, since it's designed
specifically for this kind of "run some code on a certain event"
use-case, so let's do that. We no longer override consume(), and
instead install middleware which listens for 'process_boot'. This
happens once during startup, from the main thread.

Since we can still have multiple worker processes, it's still
possible for initialization to happen from different processes
at the same time, so let's also improve the handling of this to
cut down on misleading log messages:

- use an advisory lock so that only one worker can try to deploy
  the schema at a time
- ask the DB if relevant table exists before trying to apply the
  schema, so there's no exception in the typical case
- be less verbose in logs when schema is already in place