Add missing 'async' in various places, reduce unneeded threads [RHELDST-4857]
In the case of both endpoints, and functions used with Depends, if they
don't use any blocking functionality then we should mark them as
async.

We should do this because otherwise FastAPI is pessimistic: it assumes
any non-async function might block, and so runs it in a threadpool.
This works but is rather wasteful to do this for every dependency
when not needed.