examples: start using https & certs in exodus-sync
Since we have resolved the problem with https requests hanging in
boto, we should now get into the habit of using https and client
certs/keys by default. Let's also exercise a few of the other APIs
in this example, particularly 'whoami' so the cert/key can be
easily confirmed as working (or not).

Right now, this doesn't really affect anything, but we will soon
start enforcing appropriate roles on each endpoint. Once that
happens, usage of https will become more or less mandatory.