Add state field to publishes [RHELDST-5083]
Publish states are now updated at appropriate points in the commit
process and are checked before adding items to or committing publishes.

The get_publish_by_id CRUD function was removed in favor of directly
querying as needed.

Additionally, a bug wherein delete requests to DynamoDB contained an
invalid number of keys was found and fixed.