Remove from_date Item attribute [RHELDST-4868]
This commit removes the from_date attribute from the Item model and
schema so users are unable to set this field themselves, as there is
currently no use-case for them to do so.

The from_date attribute is still required by the DynamoDB table,
however, so it is programmatically set to the datetime at which
the item's parent publish is committed.