tests: refactor db fixtures for improved usability
Within tests, when we want to access the DB it's almost always the
case that we're interested in a populated DB, where the application's
schema has already been put in place. Let's make the 'db' fixture
work that way. This allows a slight simplification of some tests
which were invoking a TestClient only for the sake of running
migrations.

In the rare case that a test wants to access the DB *before*
migrations have run, a separate fixture is provided which allows
for this.