docs: tweak language around S3 compatibility
The language here might have suggested we are only compatible with
boto (the AWS SDK for Python). In reality, the AWS SDKs for all
languages ought to work as long as they provide a way to set a
client cert and key for HTTPS, and at least the SDK for Go has
been confirmed as also compatible.

Let's keep the boto example in docs, but reword slightly to make it
clear that this is only given as an example, and just about any
existing S3 client ought to work.