Fix push-docs failing when docs haven't changed
Since we moved from Travis to Actions, the push-docs step started
to fail after merging if docs haven't changed since last push.

The behavior of the script didn't change, but it seems that Travis
simply ignored the exit code of an "after_success" command since
the job had already succeeded by that point. Actions works a bit
different and it's important to exit with a 0 exit code in this
case.