Remove various incorrect or redundant deps
Noticed while reviewing dependabot changes that various
dependencies are listed which shouldn't be, as well as some
unnecessary pinning in ".in" files.

The reason for these being added wasn't explained in the
commits which added them. Some of them are completely unused
while others are used indirectly, neither of which need to be
listed here.

In some cases, it seems they were probably to help with developing
the code on older Python versions, but we don't support anything
older than Python 3.8 now. In other cases it was probably a
misunderstanding how pip-compile works or just some guesses at
needed dependencies.