Accept publish items with link_to or object_key [RHELDST-8586]
Previously, exodus-gw accepted items that had a web_uri and a valid
object_key. Soon exodus-rsync will send items with a link_to value
instead of an object_key and the resolution of the object_key will be
handled within exodus-gw based on the link_to path. With this commit
exodus-gw accepts items that have either a link_to or an object_key.