Resolve item object_keys from link_to [RHELDST-8587]
Now that items with either object_key or link_to are accepted,
we can use link_to to find the correct object_key from among the other
publish items.

This commit adds the resolve_links function to the publish model.
If the function doesn't find an item with a web_uri that maches the
link_to of the target item, an error of code 400 is raised.