Introduce exodus-publish example/helper
This example makes it a bit easier to create and commit publish objects,
as it is somewhat cumbersome to do that otherwise (e.g. using curl + jq).
It can be used something like this:

    # Make a new publish
    PUBLISH=$(examples/exodus-publish --url ... new)

    # Do a few commands using that publish
    exodus-rsync --exodus-publish=$PUBLISH src1 exodus:/dest
    exodus-rsync --exodus-publish=$PUBLISH src2 exodus:/dest

    # Then commit it
    examples/exodus-publish commit

The idea is that, since Pub is not yet ready to manage the publish
objects for us, this script provides a reasonably convenient way to
create & commit publishes meanwhile.