tests: add a test covering common behaviors for every endpoint
There are a few invariants which all endpoints should try to satisfy in
order for the API to work correctly and consistently. Add a test which
will check these on all endpoints, to protect against mistakes when
introducing a new endpoint.

One existing test needed some minor refactoring, as it would otherwise
leave a test route installed on app which would fail test_common.