dev env: tweak sscg execution conditions
The dev env runs sscg in order to create certs.
It previously worked by always running sscg but tolerating an exit code
of 17, which was sscg's documented code for "output already exists".

Let's change it to instead only run sscg if one of the outputs doesn't
exist or is empty, because:

- the sscg "ERROR: File exists" message can be confusing and lead the
  developer to think that something has gone wrong.

- it is potentially useful to be able to recover from "cert exists but
  is empty", as this has been observed to happen at least once (though I
  don't know why).

Additionally, the duplication of sscg between exodus-gw-cert and
exodus-gw-sidecar has been removed so that this logic is only in one
place.